refactor: move all components from 'page-components' in 'app' subfolders
