+++ src/app/(main-page)/_components/MainHero.tsx
