--- src/app/catalog/[slug]/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Page() {
-  return <></>
-}
-
-export default Page
