+++ src/app/catalog/_components/ObjectDetail.tsx
