+++ src/app/catalog/_components/SimilarObjects.tsx
