+++ src/app/catalog/complexes/[complex]/_components/DetailedInfo.tsx
