+++ src/app/catalog/complexes/[complex]/_components/LayoutChoice/LayoutChoice.tsx
@@ -3,8 +3,8 @@ import Section from '@/layout/Section'
 import Container from '@/layout/Container'
 import Button from '@/ui/buttons/Button'
 import { DetailComplex } from '@/types/Complex'
-import LayoutInner from '@/page-components/complex/LayoutChoice/LayoutInner'
-import LayoutTypes from '@/page-components/complex/LayoutChoice/LayoutTypes'
+import LayoutInner from '@/app/catalog/complexes/[complex]/_components/LayoutChoice/LayoutInner'
+import LayoutTypes from '@/app/catalog/complexes/[complex]/_components/LayoutChoice/LayoutTypes'
 import { LayoutContextProvider } from './LayoutListContext'
 import ObjectsTable from './ObjectsTable'
