+++ src/app/catalog/complexes/[complex]/_components/PageSlider.tsx
