+++ src/app/catalog/new-building/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Category from '@/page-components/catalog/Category'
+import Category from '@/app/catalog/_components/Category'
 import { getNewBuildings } from '@/globals/api'
 
 async function Page() {
