+++ src/app/catalog/secondary-housing/[subcategory]/object/[object]/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ObjectPage, SubcategoryPage } from '@/types/Page'
 import { getObject } from '@/globals/api'
-import ObjectDetail from '@/page-components/catalog/ObjectDetail'
+import ObjectDetail from '@/app/catalog/_components/ObjectDetail'
 
 async function Page({ params }: ObjectPage & SubcategoryPage) {
   const object = await getObject({
