+++ src/app/catalog/secondary-housing/[subcategory]/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { getSecondaryHousingSubcategory } from '@/globals/api'
-import SecondaryCategory from '@/page-components/catalog/SecondaryCategory'
+import SecondaryCategory from '@/app/catalog/_components/SecondaryCategory'
 import { SubcategoryPage } from '@/types/Page'
 
 async function Page({ params }: SubcategoryPage) {
