+++ src/app/catalog/secondary-housing/object/[object]/page.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { ObjectPage } from '@/types/Page'
 import { getObject } from '@/globals/api'
-import ObjectDetail from '@/page-components/catalog/ObjectDetail'
+import ObjectDetail from '@/app/catalog/_components/ObjectDetail'
 import { getSimilarObjects } from '@/globals/api/methods/catalog-details/getSimilarObjects'
-import SimilarObjects from '@/page-components/catalog/SimilarObjects'
+import SimilarObjects from '@/app/catalog/_components/SimilarObjects'
 
 async function Page({ params }: ObjectPage) {
   const object = await getObject({
