+++ src/app/catalog/secondary-housing/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { getSecondaryHousing } from '@/globals/api'
-import SecondaryCategory from '@/page-components/catalog/SecondaryCategory'
+import SecondaryCategory from '@/app/catalog/_components/SecondaryCategory'
 
 async function Page() {
   const category = await getSecondaryHousing()
